+++ app/src/main/AndroidManifest.xml
@@ -12,14 +12,14 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Sharon"
+        android:theme="@style/Theme.SharonApp"
         android:networkSecurityConfig="@xml/network_security_config"
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/Theme.Sharon">
+            android:theme="@style/Theme.SharonApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
