--- app/src/main/java/com/example/sharon/ui/theme/Color.kt
@@ -1,13 +1,0 @@
-package com.example.sharon.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Gray = Color(0xFF151515)
-val White = Color(0xFFFFFFFF)
-val Ivory = Color(0xFFFFF7D1)
-val DarkViolet = Color(0xFF563A9C)
-val Violet = Color(0xFF6A42C2)
-val LightViolet = Color(0xFF8B5DFF)
-val Green = Color(0xff00dd00)
-val Yellow = Color(0xfff8d000)
-val Red = Color(0xfff00000)
