--- app/src/main/java/com/example/sharon/ui/theme/Theme.kt
@@ -1,27 +1,0 @@
-package com.example.sharon.ui.theme
-
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.darkColorScheme
-import androidx.compose.runtime.Composable
-
-private val DarkColorScheme = darkColorScheme(
-    primary = LightViolet,
-    secondary = Violet,
-    tertiary = DarkViolet,
-    background = Gray,
-    onBackground = White,
-    onPrimary = White,
-    onSecondary = White,
-    onTertiary = White
-)
-
-@Composable
-fun SharonTheme(
-    content: @Composable () -> Unit
-) {
-    MaterialTheme(
-        colorScheme = DarkColorScheme,
-        typography = Typography,
-        content = content
-    )
-}
