+++ app/src/main/java/com/example/sharonapp/screens/WaitingRoomClass.kt
@@ -1,4 +1,4 @@
-package com.example.sharon.screens
+package com.example.sharonapp.screens
 
 import android.content.res.Configuration
 import android.util.Log
@@ -40,19 +40,21 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.pointerInput
-import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.example.sharon.ui.theme.Green
-import com.example.sharon.ui.theme.Red
-import com.example.sharon.ui.theme.SharonTheme
-import com.example.sharon.ui.theme.White
+import com.example.sharonapp.ui.theme.Green
+import com.example.sharonapp.ui.theme.Red
+import com.example.sharonapp.ui.theme.White
+import com.example.sharonapp.utility.Checkconnection
+import com.example.sharonapp.utility.SecondApiService
+import com.example.sharonapp.utility.ServerResponse
+import com.example.sharonapp.utility.createApiService
+import com.example.sharonapp.utility.isRunningResponse
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.withContext
@@ -70,7 +72,7 @@ val testData = listOf(
     listOf("테스트용8", "8", "false", "true", "false"),
     listOf("테스트용9", "9", "true", "true", "false")
 )
-class WaitingRoom {
+class WaitingRoomClass {
     companion object {
         @Composable
         fun WaitingRoomScreen(
@@ -394,12 +396,3 @@ fun PlayerBox(
 
     }
 }
-/*
-@Preview(showBackground = true)
-@Composable
-fun WaitingRoomScreenPreview() {
-    SharonTheme {
-        val idInput = ""
-        WaitingRoom.WaitingRoomScreen(idInput, LocalConfiguration.current, nextScreen = {})
-    }
-}*/
